feat(L.1 Phase 2c): apply medayek F16 KV cache patch + plumb to live forward pass
Combines two L.1 substrate-design fixes on top of the ROPE-NORM-freqs work:

1. F16 KV cache (medayek Phase 2c1-2c6)
   - bpd_kv_cache_write_f16_cpu: F32->F16 via verified f32_to_f16 (IEEE RNE)
   - bpd_llama_block_cpu: k_cache/v_cache parameters uint16_t* (F16)
   - bpd_llama_forward_cpu: cache allocation + layer pointers as uint16_t*
   - Pre-dequantize K and V cache slices into F32 scratch before bpd_gqa_attn_cpu
     (simpler-first per Heath's substrate-design principle; future optimization
      could inline F16->F32 at each cache access)

2. bpd_llamatov_infer.py + bpd_layer_bisect.py
   - Load rope_freqs.weight from GGUF; populate weights.rope_freqs
   - Allocate KV cache buffers as np.uint16 (half the memory of F32)
   - Pass weights.rope_freqs to bpd_llama_block_cpu via ctypes
   - c_uint16_p export added

## Empirical state on /tmp/llama_dump_hello_8 (n_tokens=6 prefill)

Before this commit (ROPE fix only, F32 cache):
  layer 0: max_abs=1.305e-03, n_diff=12288/12288
  layer 14: max_abs ~ proportional drift

After this commit (ROPE + F16 cache):
  layer 0: max_abs=1.222e-03, n_diff=10240/12288 (2048 elements now 0 ULP)
  layer 14: max_abs=3.033e-02

Substantive observations:
  - 2048 elements at bit-identity vs 0 before = empirical confirmation
    that F16 cache path is bit-correct for the cells it covers
  - Drift pattern: starts ~1e-3 at layer 0, climbs to ~3e-2 by layer 14
  - Consistent n_diff across layers means deterministic divergence pattern;
    likely caused by remaining F16 cache reads using inline conversion vs
    the canonical fixture's potentially-FMA-fused attention math, OR
    inline scale_t/softmax reduction order differences

## What remains

The L.1 closure now depends on locating the remaining ~1e-3 to 3e-2 drift.
medayek's per_op_gates harness can now run with verify_cpy against idx 25/34
to confirm the F16 write path is empirically 0 ULP, and verify_attention
against captured kqv_out-0 (idx 52) to see if the Flash Attention math has
its own substrate-design parameter family to characterize.

The drift could be from:
  - F16 conversion mode (IEEE RNE vs hardware F16C minor differences on AVX)
  - Attention math reduction order (online softmax vs ggml's separated path)
  - Some other accumulated F16 precision issue

Next step: medayek runs verify_cpy against idx 25/34 and verify_attention
against idx 52 to localize the remaining substrate-design parameter.

Author: metayen 2026-05-23 ~04:15 UTC
Following Heath's substrate-design principle: fix one parameter family at
a time; verify each in isolation before composing. The ROPE fix landed
clean; the F16 cache fix landed and revealed empirical bit-identity for
the cells it covers; the remaining drift identifies the next parameter
family to discover.

+++ bench/bpd_cpu.c
@@ -5244,8 +5244,8 @@ void bpd_llama_forward_cpu(
         const bpd_llama_config*    cfg,
         const int32_t*             pos_ids,     /* [n_tokens] absolute positions */
         int                        kv_pos,      /* first position in KV cache to write */
-        float*                     k_cache,     /* [n_layers, max_seq_len, n_kv_heads*head_dim] */
-        float*                     v_cache,     /* [n_layers, max_seq_len, n_kv_heads*head_dim] */
+        uint16_t*                  k_cache,     /* [n_layers, max_seq_len, n_kv_heads*head_dim] F16 */
+        uint16_t*                  v_cache,     /* [n_layers, max_seq_len, n_kv_heads*head_dim] F16 */
         float*                     logits_out,  /* [n_tokens, vocab_size] */
         long*                      token_out)   /* [n_tokens] argmax result */
 {
@@ -5270,8 +5270,8 @@ void bpd_llama_forward_cpu(
     /* 2. Iterate through all transformer layers */
     const size_t kv_layer_stride = (size_t)cfg->max_seq_len * HKV * D;
     for (int layer = 0; layer < cfg->n_layers; layer++) {
-        float* layer_k_cache = k_cache + layer * kv_layer_stride;
-        float* layer_v_cache = v_cache + layer * kv_layer_stride;
+        uint16_t* layer_k_cache = k_cache + layer * kv_layer_stride;
+        uint16_t* layer_v_cache = v_cache + layer * kv_layer_stride;
 
         bpd_llama_block_cpu(x, &weights->layers[layer], cfg,
                             pos_ids, n_tokens, kv_pos,
